certbot-ci: fix integration-external tests
In 96a05d9, mypy testing was added to certbot-ci, but introduced an
undeclared dependency on acme.magic_typing, resulting in a crash when
run under the integration-external tox environment.

This change adds an explicit dependency on the typing module to
certbot-ci, and uses it in place of acme.magic_typing.